Update pid.md
I think it would be worthwhile to see the command and section in printer.cfg that is used to set the timeout values rather than hunting further to find that setting .. especially for newcomers who may have that question, rather than the old hats. The risk I guess is it becomes too big and cumbersome if every other little command is added.. but, I think in this section it is warranted. (Especially when the link to further reading has no mention of 'timeout' on the page meaning a lot more digging to find it.)